lxd/storage: Removes DaemonState() from pool interface
This same function is deprecated in containerLXC with a comment:

"This function should go away, since the abstract container interface should not be coupled with internal state details."

Which seems a valid sentiment for the storage interface too.

At the very least it shouldnt be an exported function, but as it happens its not used anyway so removing it entirely.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>